switch to 1--1 cluster-pinning redundancy
rationale: with a non -1 (= all cluster members) ipfs-cluster now do check for the remaining empty space and won't select a cluster member if the space is exausted. See changelog of v1.0.1:

Peers that are reporting freespace as 0 and which use this metric to
allocate pins, will no longer be available for allocations (they stop
broadcasting this metric). This means setting StorageMax on IPFS to 0
effectively prevents any pins from being explicitly allocated to a peer
(that is, when replication_factor != everywhere).